Change some clocks to support sub-tick precision using a new class, ClockTicks. Using that higher precision, interpolate at a higher granularity. Further, truncate the target interpolation time to vertical blank boundaries to avoid producing temporal artifacts. Fix issues caused by interpolation calculations being handled differently in multiple places (and fix cases where smoothratios were being thrown away only to be redone without checking all proper conditions). Ensure ClockTicks changes do not break other targets (EKenBuild, VoidSW), but note any interpolation there is not similarly updated.
git-svn-id: https://svn.eduke32.com/eduke32@8050 1a8010ca-5511-0410-912e-c29ae57300e0

+++ source/kenbuild/src/game.cpp
@@ -740,7 +740,7 @@ int32_t app_main(int32_t argc, char const * const * argv)
                 domovethings();
             }
         }
-        i = (totalclock-gotlastpacketclock)*(65536/(TIMERINTSPERSECOND/MOVESPERSECOND));
+        i = ((int32_t) totalclock-gotlastpacketclock)*(65536/(TIMERINTSPERSECOND/MOVESPERSECOND));
 
         drawscreen(screenpeek,i);
     }
@@ -3977,8 +3977,8 @@ void drawscreen(short snum, int dasmoothratio)
                         ptr4 = palookup[18];
                         ptr4 += (min(klabs(y2-l)>>2,31)<<8);
 
-                        j = sintable[((y2+totalclock)<<6)&2047];
-                        j += sintable[((y2-totalclock)<<7)&2047];
+                        j = sintable[((y2+(int32_t) totalclock)<<6)&2047];
+                        j += sintable[((y2-(int32_t) totalclock)<<7)&2047];
                         j >>= 14;
 
                         //ptr2 += j;
@@ -4254,7 +4254,7 @@ void drawscreen(short snum, int dasmoothratio)
 
         Bsprintf(getmessage,"Video mode: %d x %d",xdim,ydim);
         getmessageleng = Bstrlen(getmessage);
-        getmessagetimeoff = totalclock+120*5;
+        getmessagetimeoff = (int32_t) totalclock+120*5;
     }
     if (keystatus[0x57])  //F11 - brightness
     {
@@ -4320,7 +4320,7 @@ void movethings(void)
 {
     int i;
 
-    gotlastpacketclock = totalclock;
+    gotlastpacketclock = (int32_t) totalclock;
     for (i=connecthead; i>=0; i=connectpoint2[i])
     {
         copybufbyte(&ffsync[i],&baksync[movefifoend[i]][i],sizeof(input));
@@ -4528,12 +4528,12 @@ void domovethings(void)
 
     if (cameradist >= 0)
     {
-        cameradist = min(cameradist+((totalclock-cameraclock)<<10),65536);
+        cameradist = min(cameradist+(((int32_t) totalclock-cameraclock)<<10),65536);
         if (keystatus[0x52])       //0
-            cameraang -= ((totalclock-cameraclock)<<(2+(keystatus[0x2a]|keystatus[0x36])));
+            cameraang -= (((int32_t) totalclock-cameraclock)<<(2+(keystatus[0x2a]|keystatus[0x36])));
         if (keystatus[0x53])       //.
-            cameraang += ((totalclock-cameraclock)<<(2+(keystatus[0x2a]|keystatus[0x36])));
-        cameraclock = totalclock;
+            cameraang += (((int32_t) totalclock-cameraclock)<<(2+(keystatus[0x2a]|keystatus[0x36])));
+        cameraclock = (int32_t) totalclock;
     }
 
     for (i=connecthead; i>=0; i=connectpoint2[i])
@@ -4809,7 +4809,7 @@ void playback(void)
             movethings(); domovethings();
             i++;
         }
-        drawscreen(screenpeek,(totalclock-gotlastpacketclock)*(65536/(TIMERINTSPERSECOND/MOVESPERSECOND)));
+        drawscreen(screenpeek,((int32_t) totalclock-gotlastpacketclock)*(65536/(TIMERINTSPERSECOND/MOVESPERSECOND)));
 
         if (keystatus[keys[15]])
         {
@@ -5266,7 +5266,7 @@ void checkmasterslaveswitch(void)
             else
                 Bsprintf(getmessage,"Player %d (Slave)",j);
             getmessageleng = Bstrlen(getmessage);
-            getmessagetimeoff = totalclock+120;
+            getmessagetimeoff = (int32_t) totalclock+120;
 
             return;
         }
@@ -5480,7 +5480,7 @@ int loadgame(void)
 
     Bstrcpy(getmessage,"Game loaded.");
     getmessageleng = Bstrlen(getmessage);
-    getmessagetimeoff = totalclock+360+(getmessageleng<<4);
+    getmessagetimeoff = (int32_t) totalclock+360+(getmessageleng<<4);
     return 0;
 }
 
@@ -5654,7 +5654,7 @@ int savegame(void)
 
     Bstrcpy(getmessage,"Game saved.");
     getmessageleng = Bstrlen(getmessage);
-    getmessagetimeoff = totalclock+360+(getmessageleng<<4);
+    getmessagetimeoff = (int32_t) totalclock+360+(getmessageleng<<4);
     return 0;
 }
 
@@ -5720,7 +5720,7 @@ void faketimerhandler(void)
                 j = k;
             }
 
-        gotlastpacketclock = totalclock;
+        gotlastpacketclock = (int32_t) totalclock;
         return;
     }
 
@@ -5861,7 +5861,7 @@ void getpackets(void)
         case 2:
             getmessageleng = packbufleng-1;
             for (j=getmessageleng-1; j>=0; j--) getmessage[j] = packbuf[j+1];
-            getmessagetimeoff = totalclock+360+(getmessageleng<<4);
+            getmessagetimeoff = (int32_t) totalclock+360+(getmessageleng<<4);
             wsay("getstuff.wav",8192L,63L,63L); //Added 12/2004
             break;
         case 3:
